fix - member 탈퇴 시 채팅방 삭제 기능 오류 수정 및 aws 이미지 삭제 기능 추가

+++ src/main/java/com/example/server/file/AwsS3Uploader.java
@@ -13,6 +13,7 @@ import software.amazon.awssdk.services.s3.model.ObjectIdentifier;
 import software.amazon.awssdk.services.s3.model.PutObjectRequest;
 
 import java.io.IOException;
+import java.util.Objects;
 
 @Service
 @Transactional
@@ -44,19 +45,21 @@ public class AwsS3Uploader {
 
 
     public void delete(String fileUrl) {
-        ObjectIdentifier key = ObjectIdentifier.builder()
-                .key(fileUrl)
-                .build();
+        if(Objects.nonNull(fileUrl)) {
+            ObjectIdentifier key = ObjectIdentifier.builder()
+                    .key(fileUrl)
+                    .build();
 
-        Delete del = Delete.builder()
-                .objects(key)
-                .build();
+            Delete del = Delete.builder()
+                    .objects(key)
+                    .build();
 
-        DeleteObjectsRequest multiObjectDeleteRequest = DeleteObjectsRequest.builder()
-                .bucket(bucket)
-                .delete(del)
-                .build();
+            DeleteObjectsRequest multiObjectDeleteRequest = DeleteObjectsRequest.builder()
+                    .bucket(bucket)
+                    .delete(del)
+                    .build();
 
-        s3Client.deleteObjects(multiObjectDeleteRequest);
+            s3Client.deleteObjects(multiObjectDeleteRequest);
+        }
     }
 }
